fix(libvpx): fix build error
